add metal_ranger.py
When I was using JWST's exposure time calculator (https://jwst.etc.stsci.edu/) to look at the possibility of determining metallicities of distant galaxies, I had to often find which range (in microns) would be appropriate to find the metallicity of a galaxy using such-and-such calibration. This program does exactly what I want it to do, and nothing more. It's super easy to use (just python metal_ranger.py and feed it your redshift!), and easy to extend if you want to add more diagnostics.