Add kadeprimes
My partner Kade needed to choose a number for a hockey jersey. They needed three choices between 1 and 99. I realised they like primes and so we settled on 19, 37, and 73. Then a friend told them that all of these primes add to ten. This got me thinking: are there infinitely many numbers that Kade likes -- that is, primes that add to ten? Number theorists don't know! But this code lets you find them all up to the number of your choosing, and saves them all in a pretty long list.